refactor: split processor off into separate file
